Make mplayer-file-start-offset a safe local
Also add C-x SPC C-x C-f bound to mplayer-mode-find-file-at-point to the
text-mode-map. I think it's unlikely enough to clash with any other
keybinding there... and only there when this is loaded. This is required
since the mplayer-mode is a minor mode, enhancing text-mode or whatever
other mode you are in, I suppose... and so binding that function in it's
own keymap doesn't work since it's keymap isn't defined until it's
loaded. A future change could be to turn this into a regular minor mode
and have it declared in the local variables header to load it... But
this works for me right now and I'm busy using it to do high priority
work I have a deadline for.